+++ AlchemistCode/Relics/GildedKit.cs
@@ -1,27 +1,28 @@
+using BaseLib.Utils;
 using MegaCrit.Sts2.Core.Commands;
-using MegaCrit.Sts2.Core.Entities.Creatures;
 using MegaCrit.Sts2.Core.Entities.Players;
 using MegaCrit.Sts2.Core.Entities.Relics;
-using MegaCrit.Sts2.Core.Entities.RestSite;
+using MegaCrit.Sts2.Core.GameActions.Multiplayer;
 using MegaCrit.Sts2.Core.HoverTips;
-using MegaCrit.Sts2.Core.Models;
-using MegaCrit.Sts2.Core.Models.Potions;
-using MegaCrit.Sts2.Core.Rooms;
+using MegaCrit.Sts2.Core.Entities.RestSite;
+using Alchemist.AlchemistCode.Powers;
 
 namespace Alchemist.AlchemistCode.Relics;
 
 public class GildedKit : AlchemistRelic
 {
+    private const int Antitoxin = 10;
+
     public override RelicRarity Rarity => RelicRarity.Starter;
 
-    protected override IEnumerable<IHoverTip> ExtraHoverTips => new[] { AlchemistTips.Brew };
+    protected override IEnumerable<IHoverTip> ExtraHoverTips =>
+        new[] { AlchemistTips.Brew, HoverTipFactory.FromPower<AntitoxinPower>() };
 
-    public override async Task AfterPotionUsed(PotionModel potion, Creature? target)
+    public override async Task BeforeCombatStart()
     {
-        if (potion.Owner != Owner) return;
-        if (potion is FoulPotion && Owner.RunState.CurrentRoom is MerchantRoom) return;
         Flash();
-        await CreatureCmd.GainMaxHp(Owner.Creature, 1m);
+        await PowerCmd.Apply<AntitoxinPower>(new ThrowingPlayerChoiceContext(), Owner.Creature,
+            Antitoxin, Owner.Creature, null);
     }
 
     public override bool TryModifyRestSiteOptions(Player player, ICollection<RestSiteOption> options)